:wrench: turn off SDL Vsync
VSync is usually a good idea for when we don't want tearing artifacts.
The issue is that GRID doesn't have access to VSync (since our setup
can't use Protomatter and is limited by RGBmatrixPanel). So to keep
consistency, we turn off VSync. It also avoids issues where the SDL can
throw off the timing which we want to stay consistent and deterministic
like the hardware.